requirements.txt
fastapi
pdfplumber==0.11.4
python-dateutil==2.9.0.post0
gunicorn==22.0.0
uvicorn[standard]==0.30.6
